feat/#10 : IndexInfo
