+++ google-finance/src/assets/index.jsx
@@ -4,3 +4,4 @@ export { default as SubMenuIcon } from './icon/ic_menu_dots.svg';
 export { default as SearchIcon } from './icon/ic_search.svg';
 export { default as FooterLogoImage } from './img/Google.svg';
 export { default as LineImage } from './img/LineImage.svg';
+export { default as BtnRight } from './icon/ic_btn_right.svg';
