+++ google-finance/src/pages/MainDetail.jsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import styled from 'styled-components';
-import HeaderIndex from '../components/Detail/HeaderIndex';
+import HeaderIndex from '../components/Detail/HeaderIndex/HeaderIndex';
 
 const MainDetail = () => {
   return (
     <HomeWrapper>
-      주가 상세페이지 입니다.
       <HeaderIndex />
     </HomeWrapper>
   );
@@ -14,7 +13,6 @@ const MainDetail = () => {
 const HomeWrapper = styled.div`
   display: flex;
   flex-direction: column;
-  max-width: 90%;
   margin: 0 auto;
 `;
